Optimized constants.

+++ src/LoLaLinks/Abstract/AbstractLoLaLinkClient.h
@@ -15,9 +15,9 @@ private:
 	using Linked = LoLaLinkDefinition::Linked;
 
 	static constexpr uint32_t CLIENT_SLEEP_TIMEOUT_MILLIS = 30000;
-	static constexpr uint32_t CHANNEL_SEARCH_TRY_COUNT = 3;
+	static constexpr uint8_t CHANNEL_SEARCH_TRY_COUNT = 3;
 
-	static constexpr uint32_t CLIENT_AUTH_REQUEST_WAIT_TIMEOUT_MILLIS = 50;
+	static constexpr uint8_t CLIENT_AUTH_REQUEST_WAIT_TIMEOUT_MILLIS = 50;
 
 	enum WaitingStateEnum
 	{
